WifiScanner is abstracted away from the thing providing the scanning (a 'retriever'). This should better be able too support other hardware/software that provides an output that the scanner can then read.
Airport is going away; APPL deprecated it. The plan  is to use iwlist, but I cccan't dev on the mac using that. I will mock this using a file later so I can vverify the other retriever.

I may cchange this to https://packaging.python.org/en/latest/guides/creating-and-discovering-plugins/
using  either a naming  coonventioon oor a namespace package later. The goal is NOT to have the business in the __init__.py file that requires a line for the class.

For now, still working.