pulled scanner out of WIFI module and called in SCAN.
This calls the MAPAggregator for data; MAPAggregator handles either direct calls to the aggregation or redirects to the module context for 'processed' data (don't redo the work, go where it's being done).

muting adding and removing still use the module context, but this will change since the context is discrete to the module.