changed printed output and changed 'frame_id' to f_id'.
setting contours and is_inside; instead of evaluating the frame comparison on every pass, we only deal with it if the ml is outside the rect.

If the compare is still 'in range' we id this as the previous frame; ml is being shared with another contour that should not rename this one.

If the ml is inside a rect, we don't bother; this is a new thing in  a new location with a new tag.

This appears to have caught interstitial motion  without renaming immediately, but I am not convinced this is the right approach.