changed conts to contours in process_frame() to differentiate from unsorted contours. These are the unsorted ones.
process_contours() accepts these unsorted contours, sorts them and iterates over a slice of them (for now 1,2 or 3). This notably changes the printed output that each contour group is reported for each frame. Eventually, they agree on a tag, but...

The contour is used to make a wall and a rect; optionally a histogram (another story, slow...).

This contour group, wall and rect are passed to the existing tracker where each group is overlaid onto f.