added fields, methods for MSE & SSIM. playing with this as a replacement/augmentation to euclidean distance of f/wall.
Moved printed output to it's own method. Some ideas are commented.

There are two halves; labeling:
 iterate my currently tracked items.
 measure distances, find closest.
 label current based on previous name of closest.
 instrument and print something (opt.)
 append to a list
 return a list of labeled items

...and tracking:
get my mean location by iterating a list of contours in a rectangle.
 iterate a list
 Further instrumentation
 conditionally give it an id
 add to a list
 print something if it's NEW or a continuation.
 clean some cache
 return a list of 'tracked' items

This iteration oved FrameObjekt creation into 'label' locations in an effort to keep like functionality together (FINALLY). It also removes duplicated efforts.

Right now the exercise is validating the instrumented output to understand that events are faithfully reproduced. I suspect they are still not where they need to be, but this iteration is cleaned of unnecessary processing and needs commit.

I'll also likely move all the instrumentation into one, the other or it's *own* method.