fixed a bug where the capture would fail to  initialize when set to 0. This means the local video source can't be used.
The issue was setting the capture frame rate, width and height when the capture is initialized. We don't need to do this, but doing it and it being set to None is not optimum.

The solution is to *remove* the settings that do nothing for us to begin with; these add additional complexity to achieve the same result.