modifying tracking indicator behavior: red: no fix. error. yellow: offline, with fix green: fix, JS
This is based on geolocation being available.

orange: no hardware fix
purple: hardware fix

This is based on geolocation (geo) changing; we can't base hardware (HGO) position access on this because we are not using javascript. geo is unavailable because we are using H.

So, setting the TrackingButton indicator is fraught. right now it works okay, but when the page is reloaded -- it's blank. I'm thinking that map doesnt fire an onchange() event unless I literally change it -- which I don't want to do. I could move hardware out of geolocation.onchange and move it to hardwarelocation.onchange() in it's own class.

The other approach may be to make the TrackingButton indicator its own class and modify it.

Oh, and not having geo breaks the TrackingButton. It's either checking for geo to flip it or well able to control H! :/

I don't like this code. Yet.