node js_gps.cjs works, but it appears that node:http isn't available.
This may be a problem; I believe I should be able to stand up something on another port and serve it, but this may be the wrong approach.

At least I learned more about how modules work.

Other artifacts are from the rebuild. I ran 'npm init -y' on node_modules & js_gpr_ret, and this exported my function to the build as 'publisher'. I set the config to rebuild and run, hence new files.

So yeah, this would work if I can get the createServer() function to import properly.