at some point, I broke this. The issue is that the file would not be cut properly; it contained data, but was not playable as a .wav file unless transformed elsewhere. This was different behavior than expected.
The way recording was originally structured relied on a field called recording; I attempted to supplant this with the notion of a thread in cconditional statements, and this fails for a couple reasons (access [fields, methods and their scope ] and dependency [again scope]).

The bottom line is that this can not run as a daemon in it's current iteration. The  way it's being run associates this process with the parent running it, hilarity ensues.

Right now, the plan was to integrate this with cam (or reuse it in some other way). I need this to cut files correctly at minimum, so I don't care about the error.

other small changes include refactoring the order in which operations are carried out (moved file creation, refactored streamCallback(), loading config, movving file_writing_thread out of the class so it's not writing 'in class' while trying to shut down).