tracked_signals and ghost_signals are maps; parsed_signals is a list.
I can't return tracked/ghosts as JSON; these are dictionaries and because I need them to be lists for display, I have to transform the data.

That data isn't complete; it doesn't have the current values from its worker. This is updated, but IMO the source shoould have already had the things that are NOT dynamic (BSSID, SSID, Vendor,  Encryption, etc.).

Since 'ghosts' are by definition 'gone', these would not be in parsed_signals. They are 'tracked' however, and because of this  should be updated. I  do some woork  to understand whoo these ghosts are, but  this is based on a frozenset, alteration of source breaks the set.

'tracked' items should have current values, but may not depending on when requested. Because of this, update_sgnl_dynamics() seems okay.

I should be able to get data directly without static updates & transformation.

I would need to change tracked & ghosts to lists.  I may still   need to update_sgnl_dynamics().