the add() method appears to require more work than it should. I've updated it to use a new method get_cell(), but IMO this info should be updated.
The previous assertion regarding ghost being a map was incorrect; it's a list and that made me address the long stink of tracked signals being a map. now it's a list; it uses the update and update_sgnl_dynamics methods to rehydrate mappings of the bssid it contains; ditto with ghosts.

this simplifies some things; we're only dealing with lists and there is a single source of the truth for signal data. Thhe downside is that the bssid needs rehydration.

Which bring me bacck to add(); the data sshhhoould already be there -- why is SSID missing?