the add() method seems right-sized for what it does. I don't need get_cell(), because I update the  SSID field at the correcct time.
note that workers may not have been created on the initial iteration; we were looking in scanner.workers  -- but those are noot updated or even found pre initial iteration; only what iss  in the scanlist.

When signals are found they are updated; then we can look for the worker -- whic should have it's correct name by then.