I've been cheating by using __str__ to return a map; that's not what __str__ is for; I didn't know what I wanted, let alone it's type.
strings are ookay, but they don't expose the data in a way that could be readily queried; a mapping does that.

I fixed all the __str__ and moved the functionality to a method called get(); that way things are still returned to the things that need them.

I did not implement __str__ in these places; if it was there I left it, if it was close I made it work. But if there was no need for it RIGHT NOW I left it alone.

Also removed commented from StreamService