EBS lost its' controller; we don't need it, or actual routes, templates, static files or any of the REST stuff. What was in __init__.py took a hike too.
This doesn't need this stuff; as it turns out I was mistaken about what app context can provide if there is not overarching context (an idea who's time has passed) to glue it together. VIEW may be come that, but probably not as I have other plans.

The change to routes leaves only a configured EBSManager for another class to use.