added code to print cv build info if capture fails to open. I upgraded opencv python to 4.10 and it didn't use FFMPEG; the clue was that I couldn't view http streams.
This brings up another thing I've been whinging on; dependency management, the build process and packaging.

To address this, I'm starting with understanding the existing module deps; not my first rodeo. I've also ccordoned off the venv for this project as "MagPyEnvironment" (sic) in conda.

The idea right now is to get a baseline; this needs to include a couple other things I ccan think of, largely 'existing' deps like FFMPEG and the like, and 'planned' for things like pyaudio, tensorflow runtime, etc. that are deps for upcoming functionality.

Then, it's about reproducing this in a platform independent way and creating a docker image for deployment, but not getting to that now.