buttons and filters are working, but the filter only works on the 'wifi' page -- even if tracked or ghosts are being displayed.
thhe issue is caused by container being hardcoded into the filter; I need to find a way to get a handle on the tabulator.tabulator.table -- not the button.

I think I could solve this onLoad, but the it's not built yet; nothing to get.

committing until I can think of a different way to do this.