filter is still a problem, but the autoreloader works and is specific to the tab it loads.
the reload button no longer reloads the page, it reloads the data in the background.

I've noticed some caching behavior that I can attibute to the decoupled nature of what is going on. data is being cached or federated at different levels; I think I am seeing is the delay between the last MAPAggregator read of it's source, the delay of the read from the source, and the caching going on in Tabulator.

I may look into this; it's not a problem that there is a delay, and it's also not a surprise.