some notes on what  might be the problem.
this only happens here in this module, so other components have been removed as variables by established code.

I'm doing waay too much work here to start and stop. The design is still not right, but I don't want to make it monolithic as a single class.

I think this stems from some naming conventioon  that easily confuses elements of the design, and this is largely because the effort was organic. Expressly not faulting myself for not having the correct design for a problem I didn't fully understand. There have been and ccoontinue to be learnings.

This is a good stopping point; I will reevaluate the design in the near future.