completed refactoring of crop function.
the height/width are modified in the plugin what the cropper typically outputs (x, y, w, h) to allow for the way pixels are defined in openCV.

The dimensions produced by the tool are derived from the dimensions of the input image; that way multiple resolutions are supported. This differs from the previous implementation in that the class should always be getting dimensions in context to what it is cropping.

The size for the videodiv is problematic right now. It should be dynamically defined, but is currently in CSS.

I am certain there is cruft in here, but having it be *working* is more important than having it be perfect -- I need to see how it fails to understand what is next.