changed how plugin and streamservice are started.
CAMManager das a plugin (writer), a streamservice (reader) and a streamservice thread (emitter). run() call main() that instantiates a plugin reference, passes it to init() for configuration and loading.

main calls plugin.start(); the plugin initializes and starts the streamservice. It then assigns an internal thread reference to _start_plugin() which loads and processes frames.

streaming frames is done in the usual manner from the plugin implementation. The difference is that the implementation no longer reads frame 0 as a signal to start streaming.

There are still missteps in here. imshow is broken right now, likely because it or waitkey are is being called at the wrong time. I'm not certain what else, but I suspect some of the UI controls may need eval.

but this works far better that the previous incarnation; I can see it on the stream-- far less lagging going on and the stream restarts (changes direction) almost immediately.