preparation for further development.
There are places where the resolution of image doesn't need to be so high, or that I could possibly reuse fragments of representations previously made. The next iteration focuses on these areas, how they are configured and processed.

Also, there are ideas re: comparing data using vvarious metrics and distances.
I'd like to try building a test 'harness' that has all the metrics in a list and compare/ evaluate based on results to understand which works best.