cam_snap() and ImageWriter.write() now allow a frame_id to be passed. This is being used to output numbered images with symbology showing; the behavior is controlled by setting 'write_all_frames' to True.
note that setting this True and NOT passing the f_id in a call may result in overwritten frames depending on the frame rate. To mitigate this, microseconds are added to the naming convention for the file.

This formatting can be changed in config as 'DATETIME_FORMAT'. to allow for printing time, a delay of 0.2 seconds was added to the capture as 'capture_sleep_time'.