added clear_cache() method to address stale tags in cache; if > f_limit frames have passed w/o contours to analyze we clear cache.
The idea here is that 'NEW' things should not reuse tags from > f_limit frames ago.. If there has not been any motion, we need it to be a new tag.

The pla here is that now we should be able to look at the NEW location, compare it to current and know if the tag is correct.

If the new location is in range of the last tag, we know it is a continuation; reuse the tag.

if the new location isn't in range -- this is a new thing and we should create a new tag for it.