this is the most solid version so far.
it properly spots new motion (N0)
it properly spots continued motion (N1)
it properly differentiates new motion in new areas (XN)
> new motion tagged w/ new tag.
it manages the cache when no motion is detected; conceivably by then we would already have a labeled image 'shuffled' off to somewhen else.

avg_loc displays the mean location of the motion detected inside the rectangle. curr_dist displays the current delta from the previous location. If this is larger than 1x the greater of width/height of the motion, it is not 'close'.

If the avg_loc is inside the rect, we can consider this to be the correct rect of possibly multiple rects. if close is true, we can also know this is a subsequent instance of the previous motion. If hist_pass is true, we know this is a subsequent instance of the previous thing in the previous frame.

The white text displayed is wrong; it is drawn 'late' and only displays 1 thing. this could change.

after more validation on this cut, I plan to look into the gradations between t/f for the histogram (which produces a confidence 0..1, not the boolean I've mad it into), the wall (which I think needs looking into, something flaky in there), and what is being considered 'close' (_delta_pct).