not happy with this, even though it works.
The javascript that loads the cropper shouldn't be in the view_container.html.j2 template. There really is no guarantee that there will be a CAM tab, so why embed the JS there.

The other thing is maintainability; ccan't possibly know this is in here.

I may need to break up the cam.js.j2 template (the name of the file is anti-pattern (the file has html) for this project).

committing as working, but not well.