preparing to merge MAPAggregator functionality into ViewController.
by design, there is a lot of field name and functionality similarity; this shoould be a straight jump across.

I've separated out the fields so them are easy to see.
configure() will be most difficult to juggle; similar ideas.
pretty much anywhere I am asking for map.localhost I can replace with now localized logic.
run() is a straight shot across.

This will likely gut MAPAggregator and that's okay. The map is dependent on it, so it will have an anticipated lesser role so as to keep thhat module 'happy'.