fragments in VIEW templates folder should really be in the module templates folder.
fragments in the module templates folder need to be able to be included by either view_controller OR the local controller main template (*_controller.html.j2).

I am trying to create a 'convention' here; TRX now needs a template folder and a template!

this works, but I need to be able to get the template to be inccluded programatically and remove the duplicates. This is the jinja Environment loader that needs to be  addressed, the question is *when* to do this.

It may be that I need a custom loader in the container to get the  template off the filesystem.