working on load_video_new(); the issue here is that I was looading a string into the src attribute, but this code is passing a blob.
For this to work, I need to change thhe iframe into a div and set the div content to the jpeg image; that way I can use the blob and also retrieve the 'majic-color' being passed as a header in the response.

Thhe concern is that the cropper may not work if I change to a div; we will work this on it's own.