refactoring video loading; this has some things I doon't like in spite of  it's  simpliccity.
A. don't want to use iframes, I'll settle for an img, but  I think I'd prefer a div.
B. can't read the majic-color header, because this isn't a response -- it's a string being set.
C. too fast: frame retrieval is at full speed, and it needs to wait until the client consumes the frame and there is a subsequent frame ready to go.