BLOCKED on refactoring video loading:
A. iframes still offers better browser compatibility, I'll settle for an img, but I'd prefer a div.
B. still missing majic-color header on response; I am getting an initial response, but I need the subsequent ones. But since this isn't buffering in any way -- I never get to read it. So I get no  'value'. I know it is there, because I can see it printed out.
C. attempts at throttling frame retrieval require modifications to the internals of Plugin, the StreamService and its' implementation. I 'process_frame' previous to 'streaming' it, and I am streaming with my OWN service as opposed to Flask.

I'm considering using Flask as the REST service.
I'm also considering refactoring the Plugin to NOT know about that change (no more start_streamservice method in plugin -- shoouldn't be there anyway).
I will refactor the JS that loads what is now an iframe; I don't like using the iframe, but it works. I feel like it should be possible with a <div> tag, but to do so I need to be able to deal with the progressive loading (via fetch) and tag retrieval (the entire point).

I think there may be another way to get majic-color, but I don't want to go there as it is even more 'uncertain' territory compatibility-wise.