this streams properly on both channels, but leaves streamservice running on shutdown.
I suspect these can be removed if there is a BrokenPipe or ConnectionReset; I may try that next.

I could revisit this, but for now, it's more important that it work the same on both browsers and that it streams both FORE & AFT reliably. The internalized approach requires restarting the REST service 'on-the-fly' and does not replace the frame in-situ as the streamservice solution does.