unified timeouts. BlackviewGPSRetriever has issues; the thing serving is very pick about the request being the initial request fro an indeterminate amount of time (?). This equates to only the first response having the GPS data we need; from there on it's only inertia data.
I need to understand how the http connection times out at minimum, after that I could do subsequent requests? Not sure.

the other issue is that BlackviewGPSRetriever is not going to be fast enough as a network device; CAM relies on the network connection and that makes GPS data harder to receive.

The plan is to use GPSDClientRetriever, but it would have been nice to use the former. It may be the case that I can get this data between frames, but I suspect the hardware cannot support this (which is why it doesn't work).