Added beginnings of ElasticSearch integration. I've documented the docker procedures here as well as the integration plans. There are still issues to be solved with 'nested' data (signal_cache), indexes (what is here produces files that can be run to insert data -- this is NOT the plan), and how the schema is structured soo that visualizations and mapping ccan occur.
Which brings me to MAP; Elastic & Kibana could replacce the map fucctionality in a more powerful way than I have energy/understanding to build.

OSM works fine; I may still use it to implement a 'moving map' as I have come to understand that is difficult in Kibana. So it and it' symbology may still be useful.

To implement this, I have had to change thwe names of some fields. Types still need adjustment, but I did manage to get a 'geo_point' implemented for the signals.

This still needs work as per the plan, but doing this with ELK stack is anticipated to open a lot of visualization and analysis possibilities.