this will compact scanlists but, Elastic's mappings are WIP state. This is an effort to have workers on ice so that signals can be arbitrarily added later, but there are some blockers.
Time in the correct format; I am gradually refatoring how time is dealt with. The first pass was unifying field names and how they are dealt with. I created a functional method to mutate time by passing in the object [src.lib.utils.time_traveller] -- the plan is to use that and calculate those datetimes with timezone info; to satisfy the needs of both elastic and other VIEW components needing TZ aware times.