Refactoring of tracking succh as to  remove histograms and substitute SSIM  (stuctural similarity).
The primary issue with this is the outputs don't match/line up. I suspect that something is being lost or compressed here -- this doesn't appear to track in the same manner now.

I suspect this has something to do with how I am evaluating the match using SSIM.

Note that using this is just  changing the import statement for now.