sort of works; gets data from the file, fills in text_attributes and sends a completed ARXSignalPoint down the wire, sans message.
rabbit is commented out.

These don't need to be asynchronous. Having 2 assync things wait on each other in the last throes of the component life is a problem with atexit() calls. To fix, there is the internal event loop on the provider.

But, the plan now is to remove 'async' calls, including in rabbit.

I can change delivery methodss of the Provider and Consumer in the ARXMQ* classes

ARXController now has logic to start all this.

ARXRecorder should be back to it's original form.