added GPS retrieval via javascript.
righht now the index.js needs to be run from the CLI. ccd to the dir and minimally type:

node index,js

this starts the server. Loading the map page passes the current value to the server on a POST; a GET returns the most recent value.

also note; the code in  the OL distro isn't tracked. this symlink holds the actual files until they can properly be  intergrated into the vite/OL server or exist on their own.