integrated GPS retrieval via javascript in GPS module.
The controller starts the server, which is instanced in the route. Loading the map page passes the current value to the server on a POST; a GET returns the most recent value.

Finally added JSGPSRetriever block back into GPSRetriever. Note this has been plotted for a while, and this is a second incarnation.

There is a dependency on MAP running on localhost for this retriever to work. No way around that. Note: that is 'localhost', not 127.0.0.1 or anything else. For now, I'm not fixing this because it forces a bigger conversation about hostnames, resolution and DNS in this thing.

Also, be certain that Wifi is available, as this JS GPS depends on Wifi availability. It refreshes at RETRIEVER_TIMEOUT (1 second in gps.json), so it's not very granular right now.