integrated GPS retrieval via javascript in MAP module.
There is a dependency on MAP running on localhost for this retriever to work. No way around that. Note: that is 'localhost', not 127.0.0.1 or anything else. For now, I'm not fixing this because it forces a bigger conversation about hostnames, resolution and DNS in this thing.

Also, be certain that Wifi is available, as this JS GPS depends on Wifi availability. It refreshes at RETRIEVER_TIMEOUT (1 second in gps.json), so it's not very granular right now.

the validity of the enable_hardware flag is being actively debated; not ssure it's valid give the config options.