Okay, so here is the challenge:
If TRX is writing to MQ using a USB retriever, how does it read from MQ? Should it?

If TRX is reading from MQ using a TRXMQRetriever, how/why would it WRITE to MQ only to get the same information?

The use case for MQ was to move data across 'nodes' of MagPi; I might see a situation where #2 is the case, but it's not supported. Meanwhile, this class has both production and consumption, and they don't play well simultaneously.

Note that MQ should not be a dependency unless using an MQConsumer via TRXMQRetriever. This class handles production and consumption simultaineously.

TRXMQProducer has production and consumption simultaineously, but is doing it incorrectly and is duplicative effort that was factored out -- the class is unused now.