Breaking the 'test/dev' elastic integration idea out to their own components so they can be reused generically.
Mapping implies 'transformation' and this is happening on 2 levels. The structure 'onion' needs unwrapping previous to putting it to elastic, and the data needs to be restructured according to the business rules in play -- which will change and would be different for data of a differing type/structure. This is a first pass, including the name.

Meanwhile, the client class need know none of this; it's just providing the service.

the other player got these things added/removed (i.e. I'm still pushing to elastic in WifiScanner, not MapAggregator -- but I could still if needed).

This is a first pass.