refactored worker signal processing to do the same thing, in aa new order.
This ElasticMappingTransformer takes a mapTxfmr (a 'mapping transformer') on the constructor. It uses this constructor to parse the workers and signals using the mappings in the config (WORKER_INDEX_MAPPING, SIGNALS_INDEX_MAPPING).

Signals are parsed first; from this we fill self.signal_data when we process signals (this filters on worker being seen/processed before, as before.)

Then, we process the workers in the dats; in this model the only real change is the ordering of the processing and the additional feature of being able to define the transformation at compile time.
Note we could also get the transformer from the config if we had a key in the config! This could be useful to define the transformer as a configured global thing as opposed to the ad hoc call here.