shooting the trouble; I now believe the problem is with the data.
I may need to deal with signal strength to get this finally right, but the fact remains that the distances generated are too 'close' to the source to be accurate.

This crashes the minimization; it's clear from the output that it take data to get the location -- the amount I have may not do what I want in the current iteration.

The plan now is to synthesize larger scaleed data points so I can test this. I am wondering if I can simulate that through data scaling with a propagation algorithm.