updated tab name.
This is hardcoded here, but is dynamic in the viewcontroller. This component really belongs in VIEW; it doesn't make sense to create it and the copy/paste into another module.

This means there needs to be some configuration or automation to dynamically provide tnis to module. I am thinking about how this looks, it doesn't need to be much, but it should be 'clear'.