reporting works differently under MQ because of faster iterations.
Wifi not being available for scanning no longer allows len(self.scanned) > 0 to be used as a flag; MQ always produces -- even when there is nothing 'new'.

the report flag needs to be based on something else; I could:

callback to scanner: This seems like a reasonable thing, if the scccanner can't get a sccan -- perhaps it should know. what we do after that??
move the reporting: this processing is going to be factored out to another class later, perhaps this is an opportunity to extract this too.
not have the check at all: If scanning is failing to get anything, something is misconfigured and we should stop: "Wifi isn't on".

Not sure. This isn't on the path -- leaving it for now.