rolled thiss back to pre-mq.
I want the aggregator to use the module retriever when getting data. If there is no data from the module, but other data from other modules -- with the current incomplete code the wrong data and goto the wrong module.

The idea is a 'mapping' of retrievers when registering modules. When aggregate runs, iot cycles through the mapping and eacch modules implementation of getting data.

It's becoming over-engineered for what I need right now. REST works fine, but I would *rather* use MQ for network comms between devices.

I'm going to table this for a bit until the MQ implementation matures; meanwhile -- it's using REST and still working.