First pass at SDR (what I am calling it now); targeting:
A continuous spectrogram from stick (based on pyrtlsdr demo_waterfall.py as base): this code works (it's ), but the UI cannot be MPL. I did research on jupyter that I'll pull in later. The point is to get some data from a toy SDR to play with.

> find frequencies in the spectrogram that fit a 'model' (avg db, snr, on, etc...): This would be a 'window' of sorts, likely based on signal strength, like being within a range tf.
> assign a 'worker' to monitor the frequencies that 'fit': This means 'SDRWorker':
- records: convert I/Q data to audio arrays.
- catalogs: creates and fills SDRSignalPoint with audio data and metadata.
Also:
- all signals are displayed in the spectrogram UI. signals with workers are highlighted in the UI as a shaded/box travelling with audio.
- there are plans for UI elements below the spectrogram that manipulate the tracking/tasking on signal.

There is a minimal UI diagram in the stacks. There are also 'control fields' in the signalpoint.