refactored aggregation; this appears to work, but neglects needed processing.
The interface to retrievers wasn't created because I'm not solid on what needs 'retrieving'. I can see a fair amount of room for improvement now that we have some actual retrievvers to play with.

This reads the config and loads the proper retriever for each live module passed ONLY WHEN MQ IS ONLINE. Otherwise, it uses REST.

has MQ?
> Y: load retriever.
-> If it loaded, use.
> N: don't load retriever.
-> No retriever? use REST
Error?
use REST
Error?
log the error.

Everything else is the same as it ever was.