thia is fine for now and it works, but it raises questions about the design. The real issue is that I haven't settleed on a transport mechanism; I have REST and several flavors of MQ to play with, and I want to use each with specificity.
'stats' are getting used here and there, but nothing like what's planned. 'stats' are being treated like 'extra kids' in this thing because I'm to engaged on other higher priority issues (function).

I pull stats as needed; and this is how things suffer.

even further back, this is a shadow of MAPAggregator, which is specialized for MAPS and ONLY ever supposed to use REST.

And, it's complicated. It's likely slowing me down and I don't know it. I'm leaving it for now; it's not affecting function.

I need to make some choices.