tyring to implement a check for ZMQ, but having trouble due to address being in use.
I am doing this wrong. I don't want to load MQProvider too do the test; I want to know is MQ is working at a baseline. MQProvvider is specific to ARX and should stay that way -- I want to split the metadata out to rabbit.

The plan is look @ ZMQ unit 'tests' as a model.