made the decision to remove RabbitMQ.
My ideas are good, but I'm doing this wrong. The way this is structured now is complicated because I wanted to try some things out. I've since decided which flavor is of most benefit, least overhead and doesn't need a lot of extra had-holding to bee actionable. ZeroMQ fits the bill; it'ss lightweight and flexes well; RabbitMq needs a server startup and while it's entirely possible to use it -- needs a lot more management than the others.

Did I mention that ZeroMQ & ImageMQ are related? ImageMQ uses ZeroMQ under the hood -- If I standardize on Zero for everything I can devvote ImageMQ to CAM, further simplifying things.

I believe I made these experiments in a way that splits data from metadata and passes it over the apropos wire. This CAMMQProvider has removed RMQ. There is an issue causing the connection to block, but that's not a priority right now.