Refactoring ZeroMQ modules; I've made progress on what patterns get used where, discovering the error of my ways.
I mostly need PUSH/PULL for ARX (only on stop() call) and WIFI (passing data from retriever). TRX needs to intermittently connect (pick up spurious transmissions, evaluate broadcast periodicity and source) so REQ/REP for that and SDR needs multiple workers to connect to  a single server intermittently.

I will need to be able configure hosts and ports. socket type is what the classes are for, bind vs. connect is part of the pattern, so I may find that I abstract that out later.

WHat I don't want to do is copy/paste this code everywhere as a one-off.