this is working, but there is a naming conflict with 'scanner' in trx.
I need to sort that, but am more concerned with data and how it is displayed. There are ugly bits all over the place; I have not been concerned with keeping UI consistent because it's changing based on the model.

There  are columns texxt_attributes that need handling and manufacured columns that need promulgation, but only for some things and right now I HAVE NO IDEA WHAT THOSE ARE.

So. I commit and move on.