I have a better idea; I'm thinking to record the entire bandwidth of frequency domain information in aggregrate and mine that for signals. When mining, I can find the peaks in the newly recorded ƒ domain data. From the peaks I can assign worker processes to tske that slice of frequency, ccoonvert it to † domain and monitor it.
This way, I only need ONE piece of hardware to record signals. I ccan have unlimited number of 'playback' heads, and within the recorded bandwidth -- I'd never miss a signal.

This would be a great place to the SSET SignalFrame and TimeFrequencyFrame. Those componeents are designed to deal with †/ƒ conveersions and slicing.
SDRSignalPoint doesn't do this, and it would be more valuable at this point to try using SSET rather that let it eexist oon it's own.

This is also an opportunity to use the MCP integration. Having a frequency scanner under AI control seems like it might be a fun tool to have, and having AI do the searching via prompt would be *amazing*.

So, the iteration pattern for SDR is changing. This commit  passes data and creates a signalpoint, though there is not much 'separation' in the peaks (I will add analyzer to tweak this into a proper peak later) and theus workers, the file doesn't cut (not surprising), and some of the calculations might be wrong.

I have not added SignalFrame, TimeFrequencyFrame yet...