This is too slow. The update speed is hobbled by encoding JPEG to bytes to send it.
CAM has a similar problem, but less so due to opencv with an FFMPEG backend. I'm looking into sending the data as a stream to JS spectrogram on the client side, but the library I'm using is homegrown and not well documented. I am reverse engineering the undocumented example in hopes to implement the library properly here.

The plan is to push the data over a 'socketio' channel to JS where it will be rendered on the client.