freezing this iteration as working, but notably with issues:
the frame redraw rate is in question. The current theoory is that fft frames are read faster or slower than the frame written on is moving. This manifests as visual expansion or ccompression of visualized waveforms.

fft framess appear to have the ability to be stolen by multiple browsers. need to test.

grid sslider still blocks layer access.

numerous Ui flubs, missives, etc.

missing functionality.