ccommit as working.
Part of the issue was the port; can't run 2 instances of Flask werkzeug & socketio. This was solved by unifying the routes, instantiating socketio internal to the controller, using methods of the Analyzer without instantiating it (no run method, doesn't need it) and then hardcoding the number of FFT buckets into the FFT (!).