FINAL COMMIT:
This project is incomplete demo code, and won't be completed now. All these modules work on some level. It was intended as 'starting point' for monitoring and instrumenting some electromagnetic signals, and then dealing the information produced by them in intuitive and useful ways. It was an excuse to let my imagination run wild -- in Python.

There is unfinished code. I tried some thing that didn't work well, and left some things unimplemented (SDR buttons, CAM 'trapping', 'decrypting', 'blocking', etc.) as dangerous or foolish to try. Still others are ideas in the process of being implemented or deprecated in favor of other methods (REST vs. MQ). The code is disorganized, things need removing, condensing, refactoring or explaining. Still other code is working, but work was being done elsewhere at the time that was of more value. The focus is experimentation and understanding how to work with these types of signals, having programmatic control of them.

BACKSTORY: I know what this thing ended up as, and I know why I built it. I am certain there are people 'in the know' who might opine that code 'like this' could be weaponized to monitor individuals/infrastructure by the signals that they emanate. They would be correct, and that was the 'driving force' the project. My question was "How much code, and what types of code are needed to achieve a feature objective in this space?", and "Given different purposes and forms of signals, what can be derived from electromagnetic spectra? (See the later effort 'Signal Semiotics' which is a direct outgrowth of this type of rationalization)". Historically, I've used Open Source information to understand 'opaque concerns' about the world around me. So, I am fully aware of the 'misuse potential' here.

I believe in privacy and personal sovereignty. My concern is that in my personal experience over the last 10 years I have witnessed AI as 'misused' right now by online services and providers I entrusted with my own personal information. LinkedIn has been eating data for years; and it now has opinions about me ('ratings', 'likes', etc.) that I can't change, that have  markedly and negatively impacted my employment opportunity. Workday is exactly the same; their federated system requires NEW authentication for each company; this has a side effect of allowing data collection across sectors while limiting the perceived scope of the information retained.

Workday is currently being sued on these merits; I'd expect LinkedIn would be next. Consider the side effects of these recruiting practices; companiies are 'digital doxxing' and 'blacklisting' targets, thereby economically marginalizing individuals and organizations -- I believe in an effort to create an extended class of 'wage slave' labor to support a new 'Authoritarian Oligarchy' within the next 4 years.

George Orwell warned us about this; "Animal Farm" and "1984" were not reading materials for the recent crop of technologists. They have been raised to realize the monetary benefits of the services and infrastruccture they build and not the 'empathetic utility' of the thing. All services designed to 'suck a user in' to paying said service either monetarily or with 'engagement' -- the 'engagement' being a 'fight' online that everyone purportedly wants to engage on -- and this drives revvenue and this is OKAY.

I saw all this coming after the general release of the smartphone. Orwell was right, and I've seen examples all around me. Consumption of misinformation, malinformation and disinformation (along with Truth) in the firehose manner it is being presented, sans logic, common sense or empathy is feeding us all directly into an emotional wood chipper.

We have all witnessed the same technology used in misguided ways to violate privacy, target groups, and change Nations -- bu for a loong time we couldn't explain it. I was on FaceBook the eve of the 2016 election; I noticed the surge in political misinformation designed to sway opinion on that night -- so much that I posted my belief that FaceBook was being hacked. Over the next years I made it very public how I felt about FaceBook, yet nobody believed it  -- until they did.  I observed the fallout and the investigation and sadly, I was correct about more than I had hoped.

In 2021, I came to the realization that deepfake technology had reached a level of realism that human psychology and emotion doesn't fully understand how to digest. Not that it would be perceived as either genuine or artificial, but that the human social cues and mannerisms could be manipulated in an automated way so as to increase the probability that a 'fake' would be accepted as 'real'. The accepted facsimile of what we know as 'emotions' can be instrumented, programmed and weaponized, and to my knowledge -- this hasn't previously been feasibleexpressly in an interactive feedback context.

It's my belief that 'Lies' are the heart of all things we know as Evil. I am expressly NOT against the technology. I am certain there are many useful and proper applications. This system, and systems like them -- have the potential to generate 'lies' we don't understand how to deal with -- on an automated basis. Understanding all I had witnessed with 'FaceBook', Cambridge Analytica and how the ability to read each others diaries may not have been a great idea, knowing how my own life and career have been disaffected by the promulgation of information about me -- I asked myself "how will society 'beat this plowshare into a sword' and misuse it?"

All that was also on the cusp of ChatGPT. I stood by and watched the world become fascinated with, and *race* to use Large Language Models in an effort to save itself from 'thinking' only to become frustrated based on lack pf understanding and rigor. I chose to attempt identify the ways in which the technology could be misused, reverse-engineer ways to do it, and monitor the environment for systems and policies that I believe might be enabling that misuse.

I don't really expect this to ever be read; if it is -- it's gonna be misunderstood. But, this is what I see: a world that doesn't fully understand the technology that is being foisted on it. A adversarial technology that is so able to analyze, judge and subsequently destroy so many lives without empathy or context. And, it won;t do this on it's own --  it will be 'tricked' or otherwise manipulated to do so BY HUMANS -- with extreme prejudice.

White supremacy has been using the 'latest technology' to further it's racist agenda since before "The Birth of a Nation" (moving pictures), shortwave (unfettered, global reach), and now 'internet' -- the effort continues to this day. AI is the latest tool that can be used to seek and destroy individual by 'dealing in information' ("Mona Lisa Overdrive", "Snow Crash" and the 'Metaverse'). I see history repeating itself, and I see all my worst fears about information techology coming true -- and I don't care for it.

Don't use AI or machine learning maliciously. Don't use what I created maliciously either; I understand human nature -- this thing can get someone hurt. It is closer to a 'sword' than a 'plowshare'; but its' also a 'blunt' instrument that uses 'toy' data and I'm never going to finish it. In the context of someone that alters code, understands what they see, and knows how to analyze and interpret the data -- it might be an avenue to something else entirely.

 Instead, check out the conversation and ideas in SSET. ChatGPT and I collaborated  on the idea, and I think it has legs. For example, there was a discussion of automated DSL creation in an effort to exploit the linguistic bandwidth and understanding of Large Language Models, with which I make comparative observations in various taxa wrt to how their languages may have evolved.